Correct handling of navigational status in RMC and GNS
There was a misunderstanding on what the "navigation status" field
should include. The enum values were for the "positioning system mode
indicator" which is the preceding field. This changes the enum values to
correspond to actual ones.

I realize that in principle this is a breaking change, however there's
no such thing (as far as I can tell) as the previous navigational status
values (as those are in fact mode values, which we don't keep an enum
for), so I doubt anyone should depend on them... If they do, that will
be a bug by itself and breaking it to highlight the error may be
desirable.

Fixes #107